requirements.txt
tensorflow; sys_platform == 'linux'
tensorflow-macos; sys_platform == 'darwin'
numpy
matplotlib
pillow

flask

streamlit
requests
